Fix all Vader tests and simplify test runner infrastructure
- Fix autopep8.vader tests (8/8 passing)
  * Initialize Python paths before loading autoload files in setup.vim
  * Make code_check import lazy in autoload/pymode/lint.vim
  * Ensures Python modules are available when autoload functions execute

- Fix commands.vader PymodeLintAuto test (7/7 passing)
  * Same root cause as autopep8 - Python path initialization
  * All command tests now passing

- Simplify test runner infrastructure
  * Rename dual_test_runner.py -> run_tests.py (no longer dual)
  * Rename run-vader-tests.sh -> run_tests.sh
  * Remove legacy test support (all migrated to Vader)
  * Update all references and documentation

- Update TEST_FAILURES.md
  * Document all fixes applied
  * Mark all test suites as passing (8/8)

All 8 Vader test suites now passing:
✅ autopep8.vader - 8/8 tests
✅ commands.vader - 7/7 tests
✅ folding.vader - All tests
✅ lint.vader - All tests
✅ motion.vader - All tests
✅ rope.vader - All tests
✅ simple.vader - All tests
✅ textobjects.vader - All tests
