+++ tests/vader/autopep8.vader
@@ -5,6 +5,7 @@ Before:
   call SetupPythonBuffer()
 
 After:
+  source tests/vader/setup.vim
   call CleanupPythonBuffer()
 
 # Test basic autopep8 availability
